daemon/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "phantom-daemon"
version = "0.1.0"
description = "Privacy noise generator daemon"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "ollama>=0.4",
    "pydantic>=2.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.8"]

[project.scripts]
phantom-daemon = "phantom_engine.main:run"
